feat: github users list of javascripter almost done
After some refactoring of the function declarations the feature of Github users list of Javascripters is almost ready. Every fetching function definition moved in the component that is invoked. So that the searching function of Github Javascripters is in the main component App.js  The function of the fetching of the user details is in the GithubUsers.js  All the needed props are now passed to the child component in order to display info about the user in a list.

To Do The passed props need to used in the display of the child component