feat: github user list of javascript
The feature that lists Github users, Javascripters is ready. On start of the app the fetching data function is called and return a list of users with order by most followers. The structure of the React app to implement this feature is going from the App.js to the Githubuser component and then to the ProfiMedObject child component. The child component shows all the info for the user.

Pagination must be done.