pagination of the github users list by number
The pagination of the Github users list is functional by clicking on the number of the page. So that, when another button number is clicked the list updated, by fetching this page users from the API, and displaying this list of users. The technique used is the simple on-click-event to pass the handler with the id number of the html element.

Note to remember that the setState() could be queued the tasks and the fetch() method is anynced, so that dificult to perform an operation with one after the other.