feat: next and previous page navigation for the list of users
This commit implements the navigation functionality of the app, so that the app user to go to the next list of github users by clicking the Next page button, and to go to the previous list of github users by clicking the Previous button. The Previous button is disabled when the app shows the first page list of users.

The functionality of navigating directly at the first page, at the second page, and at the third page of github list of users by clicking the appropriate button 1,2 and 3 have already implimented at another commit.